Encode characters with custom functions and not using Jason's
